chore: add helper functions for flattening/unflattening public inputs
