+++ compiler/integration-tests/test/node/smart_contract_verifier.test.ts
@@ -7,7 +7,7 @@ import toml from 'toml';
 
 import { compile, init_log_level as compilerLogLevel } from '@noir-lang/noir_wasm';
 import { Noir } from '@noir-lang/noir_js';
-import { BarretenbergBackend } from '@noir-lang/backend_barretenberg';
+import { BarretenbergBackend, flattenPublicInputs } from '@noir-lang/backend_barretenberg';
 
 compilerLogLevel('INFO');
 
@@ -59,44 +59,8 @@ test_cases.forEach((testInfo) => {
 
     const contract = await ethers.deployContract(testInfo.compiled, [], {});
 
-    const publicInputIndices = [...proofData.publicInputs.keys()].sort();
-    const flattenedPublicInputs = publicInputIndices.map((index) =>
-      hexToUint8Array(proofData.publicInputs.get(index) as string),
-    );
-    const publicInputsConcatenated = flattenUint8Arrays(flattenedPublicInputs);
-
-    const result = await contract.verify(proofData.proof, publicInputsConcatenated);
+    const result = await contract.verify(proofData.proof, flattenPublicInputs(proofData.publicInputs));
 
     expect(result).to.be.true;
   });
 });
-
-function flattenUint8Arrays(arrays: Uint8Array[]): Uint8Array {
-  const totalLength = arrays.reduce((acc, val) => acc + val.length, 0);
-  const result = new Uint8Array(totalLength);
-
-  let offset = 0;
-  for (const arr of arrays) {
-    result.set(arr, offset);
-    offset += arr.length;
-  }
-
-  return result;
-}
-
-function hexToUint8Array(hex: string): Uint8Array {
-  const sanitised_hex = BigInt(hex).toString(16).padStart(64, '0');
-
-  const len = sanitised_hex.length / 2;
-  const u8 = new Uint8Array(len);
-
-  let i = 0;
-  let j = 0;
-  while (i < len) {
-    u8[i] = parseInt(sanitised_hex.slice(j, j + 2), 16);
-    i += 1;
-    j += 2;
-  }
-
-  return u8;
-}
